# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Water Station Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export type SessionUserRole = typeof SessionUserRole[keyof typeof SessionUserRole];


export const SessionUserRole = {
  station: 'station',
  admin: 'admin',
} as const;

export interface SessionUser {
  id: number;
  username: string;
  name: string;
  role: SessionUserRole;
  isActive: boolean;
}

export interface ChangePasswordInput {
  currentPassword: string;
  /** @minLength 4 */
  newPassword: string;
}

export interface ChangeNameInput {
  /** @minLength 1 */
  name: string;
}

export interface Station {
  id: number;
  username: string;
  name: string;
  isActive: boolean;
  createdAt: string;
}

export interface StationAdmin {
  id: number;
  username: string;
  name: string;
  isActive: boolean;
  createdAt: string;
  totalSalesToday: number;
  /** @nullable */
  lastLoginAt: string | null;
}

export interface StationToggleInput {
  isActive: boolean;
}

export interface Product {
  id: number;
  name: string;
  price: number;
  stock: number;
  stationId: number;
  createdAt: string;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @minimum 0 */
  price: number;
  /** @minimum 0 */
  stock: number;
}

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minimum 0 */
  price?: number;
  /** @minimum 0 */
  stock?: number;
}

export interface Sale {
  id: number;
  stationId: number;
  /** @nullable */
  productId: number | null;
  productName: string;
  quantity: number;
  unitPrice: number;
  totalPrice: number;
  isCoupon: boolean;
  createdAt: string;
}

export interface SaleInput {
  productId: number;
  /** @minimum 1 */
  quantity: number;
  isCoupon: boolean;
}

export interface DailySummary {
  totalRevenue: number;
  totalSales: number;
  couponCount: number;
  waterBottleCount: number;
  date: string;
}

export interface DailyRevenue {
  date: string;
  revenue: number;
}

export interface TopProduct {
  name: string;
  quantity: number;
  revenue: number;
}

export interface SalesStats {
  period: string;
  dailyRevenue: DailyRevenue[];
  totalRevenue: number;
  topProducts: TopProduct[];
}

export interface Expense {
  id: number;
  stationId: number;
  description: string;
  amount: number;
  category: string;
  createdAt: string;
}

export interface ExpenseInput {
  /** @minLength 1 */
  description: string;
  /** @minimum 0 */
  amount: number;
  category: string;
}

export interface Shift {
  id: number;
  stationId: number;
  workerName: string;
  startTime: string;
  /** @nullable */
  endTime?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface ShiftOrNull {
  shift?: Shift | null;
}

export interface ShiftInput {
  /** @minLength 1 */
  workerName: string;
  startTime: string;
  notes?: string;
}

export type ListSalesParams = {
date?: string;
};

export type GetSalesStatsParams = {
period?: GetSalesStatsPeriod;
};

export type GetSalesStatsPeriod = typeof GetSalesStatsPeriod[keyof typeof GetSalesStatsPeriod];


export const GetSalesStatsPeriod = {
  week: 'week',
  month: 'month',
  year: 'year',
} as const;

export type ListExpensesParams = {
month?: string;
};

export type ListShiftsParams = {
date?: string;
};