# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Water Station Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login (station or admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionUser"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current session info
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionUser"
        "401":
          description: Not authenticated
  /auth/change-password:
    patch:
      operationId: changePassword
      tags: [auth]
      summary: Change station password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
        "400":
          description: Invalid input
  /auth/change-name:
    patch:
      operationId: changeStationName
      tags: [auth]
      summary: Change station display name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeNameInput"
      responses:
        "200":
          description: Name changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Station"

  # Products
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products for current station
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Sales
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expenses for current station
      parameters:
        - name: month
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Record an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Shifts
...
  /admin/stations/{id}/toggle:
    patch:
      operationId: adminToggleStation
      tags: [admin]
      summary: Admin - enable or disable a station
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StationToggleInput"
      responses:
        "200":
          description: Station updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StationAdmin"
components:
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
    SessionUser:
      type: object
      required: [id, username, name, role, isActive]
      properties:
        id: { type: integer }
        username: { type: string }
        name: { type: string }
        role: { type: string, enum: [station, admin] }
        isActive: { type: boolean }
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string }
        newPassword: { type: string, minLength: 4 }
    ChangeNameInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
    Station:
      type: object
      required: [id, username, name, isActive, createdAt]
      properties:
        id: { type: integer }
        username: { type: string }
        name: { type: string }
        isActive: { type: boolean }
        createdAt: { type: string }
    StationAdmin:
      type: object
      required: [id, username, name, isActive, createdAt, totalSalesToday, lastLoginAt]
      properties:
        id: { type: integer }
        username: { type: string }
        name: { type: string }
        isActive: { type: boolean }
        createdAt: { type: string }
        totalSalesToday: { type: number }
        lastLoginAt: { type: ["string", "null"] }
    StationToggleInput:
      type: object
      required: [isActive]
      properties:
        isActive: { type: boolean }
    Product:
      type: object
      required: [id, name, price, stock, stationId, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        price: { type: number }
        stock: { type: integer }
        stationId: { type: integer }
        createdAt: { type: string }
    ProductInput:
      type: object
      required: [name, price, stock]
      properties:
        name: { type: string, minLength: 1 }
        price: { type: number, minimum: 0 }
        stock: { type: integer, minimum: 0 }
    ProductUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        price: { type: number, minimum: 0 }
        stock: { type: integer, minimum: 0 }
...
    Expense:
      type: object
      required: [id, stationId, description, amount, category, createdAt]
      properties:
        id: { type: integer }
        stationId: { type: integer }
        description: { type: string }
        amount: { type: number }
        category: { type: string }
        createdAt: { type: string }
    ExpenseInput:
      type: object
      required: [description, amount, category]
      properties:
        description: { type: string, minLength: 1 }
        amount: { type: number, minimum: 0 }
        category: { type: string }
...