# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Water Station Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Expense {
  id: number;
  stationId: number;
  description: string;
  amount: number;
  category: string;
  createdAt: string;
}