# lib/api-zod/src/generated/types/sessionUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Water Station Management API
 * OpenAPI spec version: 0.1.0
 */
import type { SessionUserRole } from './sessionUserRole';

export interface SessionUser {
  id: number;
  username: string;
  name: string;
  role: SessionUserRole;
  isActive: boolean;
}